+++ lib/src/models/models.dart
@@ -1,11 +1,11 @@
 export 'alignment.dart';
 export 'child_config.dart';
 export 'done_config.dart';
-export 'instruction_result.dart';
 export 'introduce_config.dart';
 export 'next_config.dart';
 export 'pre_dialog_button_type.dart';
 export 'pre_dialog_config.dart';
 export 'previous_config.dart';
 export 'skip_config.dart';
+export 'tour_action.dart';
 export 'tour_state.dart';