+++ lib/src/models/tour_action.dart
@@ -1,5 +1,12 @@
-/// The result of the `introduce` method.
-enum IntroduceResult {
+/// A result returned by a tour instruction, indicating the action taken during the tour.
+@Deprecated('Use TourAction instead of IntroduceResult')
+typedef IntroduceResult = TourAction;
+
+/// The action taken during the tour.
+enum TourAction {
+  /// The introduction is about to be shown.
+  introduce,
+
   /// [disabled] is returned when the current widget is disabled.
   disabled,